muen: Extend muen networking script
Create additional muennet interfaces if the kernel command line contains
a 'unikernel_iface' parameter. The parameter value is interpreted as a
comma-separated list of network interfaces belonging to MirageOS
unikernels. The name of the reader and writer channels are derived from
the interface name by appending the "|in" and "|out" suffixes. The
interfaces are attached to the br0 bridge which connects the unikernels
with the outside world.